fixup! move annotations, kongstate, util, parser into /pkg/
