+++ cli/ingress-controller/flags.go
@@ -27,7 +27,7 @@ import (
 
 	apiv1 "k8s.io/api/core/v1"
 
-	"github.com/kong/kubernetes-ingress-controller/internal/ingress/annotations"
+	"github.com/kong/kubernetes-ingress-controller/pkg/annotations"
 )
 
 const (
